+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [11.2.0] - 2024-03-17
+
 ### Added
 
 - Ability to create items
@@ -133,7 +135,9 @@ Dummy version to see if it fix the non apparition on the FoundryVTT list
 
 - First system version
 
-[Unreleased]: https://github.com/DjLeChuck/foundryvtt-system-channel-fear/compare/11.1.1...HEAD
+[Unreleased]: https://github.com/DjLeChuck/foundryvtt-system-channel-fear/compare/11.2.0...HEAD
+
+[11.2.0]: https://github.com/DjLeChuck/foundryvtt-system-channel-fear/compare/11.1.1...11.2.0
 
 [11.1.1]: https://github.com/DjLeChuck/foundryvtt-system-channel-fear/compare/11.1.0...11.1.1